feat: render event preview for nostr:nevent mentions
Changes:
- extracted NoteContent component from Content component. It is
meant to render only immediate content of the note (no URL previews or
other note mentions)
- changed Content component to render only npub and nprofile mentions (using NoteContent component)
- changed Content component to render nostr:nevent mentions (if any) separately after the main content.
For that we get referenced event, parse its content and use NoteContent component to render it. Also
using special "compact" view for UserProfile component to render the header for referenced event.
- to make NoteContent work we extracted content parsing logic from Content component into ParsingService.
This way we can reuse it to parse content of both event and referenced events.
- moved 'createNote' method from LayoutService into EventService to avoid fix circular dependency
