+++ src/app/components/content/content.component.html
@@ -1,62 +1,39 @@
 <div class="content-container" #contentContainer>
   @if (shouldShowContent()) {
-    @for (token of contentTokens(); track token.id) {
-      @if (token.type === 'text') {
-        <span>{{ token.content }}</span>
-      } @else if (token.type === 'linebreak') {
-        <br />
-      } @else if (token.type === 'emoji') {
-        <span class="emoji-token" [title]="token.content">{{
-          token.emoji
-        }}</span>
-      } @else if (token.type === 'nostr-mention') {
-        <span class="nostr-mention" (click)="onNostrMentionClick(token)">
-          &#64;{{ token.nostrData?.displayName }}
-        </span>
-      } @else if (token.type === 'url') {
-        <a [href]="token.content" target="_blank" rel="noopener noreferrer">{{
-          token.content
-        }}</a>
-      } @else if (token.type === 'youtube') {
-        <div class="media-container">
-          <iframe
-            width="560"
-            height="315"
-            [src]="token.processedUrl"
-            frameborder="0"
-            allow="accelerometer; autoplay; clipboard-write; encrypted-media; gyroscope; picture-in-picture"
-            allowfullscreen
+    <app-note-content [contentTokens]="contentTokens()" />
+
+    @for (mention of eventMentions(); track mention.event.event.id) {
+      <mat-card>
+        <mat-card-header>
+          <app-user-profile
+            [pubkey]="mention.event.event.pubkey"
+            view="compact"
           >
-          </iframe>
-        </div>
-      } @else if (token.type === 'image') {
-        <div class="media-container">
-          <img
-            [src]="token.content"
-            alt="Content image"
-            loading="lazy"
-            (click)="openImageDialog(token.content)"
-            class="clickable-image"
-          />
-        </div>
-      } @else if (token.type === 'audio') {
-        <div class="media-container">
-          <audio controls>
-            <source [src]="token.content" />
-            Your browser does not support the audio element.
-          </audio>
-        </div>
-      } @else if (token.type === 'video') {
-        <div class="media-container">
-          <video controls>
-            <source
-              [src]="token.content"
-              type="video/{{ getVideoType(token.content) }}"
-            />
-            Your browser does not support the video element.
-          </video>
-        </div>
-      }
+            <a
+              class="date-link"
+              [matTooltip]="
+                mention.event.event.created_at * 1000 | date: 'medium'
+              "
+              matTooltipPosition="below"
+              tabindex="0"
+              role="button"
+              (click)="
+                layoutService.openEvent(
+                  mention.event.event.id,
+                  mention.event.event
+                )
+              "
+            >
+              {{ mention.event.event.created_at | ago }}
+            </a>
+          </app-user-profile>
+        </mat-card-header>
+        <mat-card-content>
+          <div class="content-container">
+            <app-note-content [contentTokens]="mention.contentTokens" />
+          </div>
+        </mat-card-content>
+      </mat-card>
     }
 
     @for (preview of socialPreviews(); track preview.url) {
