+++ src/app/components/content/content.component.ts
@@ -14,43 +14,19 @@ import {
 
 import { MatCardModule } from '@angular/material/card';
 import { MatProgressSpinnerModule } from '@angular/material/progress-spinner';
-import { MatDialogModule } from '@angular/material/dialog';
-import { DomSanitizer, SafeResourceUrl } from '@angular/platform-browser';
 import { SocialPreviewComponent } from '../social-preview/social-preview.component';
-import { MatDialog } from '@angular/material/dialog';
-import { ImageDialogComponent } from '../image-dialog/image-dialog.component';
 import { SettingsService } from '../../services/settings.service';
 import { UtilitiesService } from '../../services/utilities.service';
 import { Router } from '@angular/router';
-import { MediaPlayerService } from '../../services/media-player.service';
-import { ParsingService } from '../../services/parsing.service';
-
-interface NostrData {
-  type: string;
-  // Narrow known shapes; keep flexible with index signature for unseen fields
-  // eslint-disable-next-line @typescript-eslint/no-explicit-any
-  [k: string]: any;
-  displayName: string;
-  data: unknown;
-}
-
-interface ContentToken {
-  id: number;
-  type:
-    | 'text'
-    | 'url'
-    | 'youtube'
-    | 'image'
-    | 'audio'
-    | 'video'
-    | 'linebreak'
-    | 'nostr-mention'
-    | 'emoji';
-  content: string;
-  nostrData?: NostrData;
-  emoji?: string;
-  processedUrl?: SafeResourceUrl; // For YouTube embed URLs that are pre-processed
-}
+import { ContentToken, ParsingService } from '../../services/parsing.service';
+import { NoteContentComponent } from './note-content/note-content.component';
+import { DataService } from '../../services/data.service';
+import { NostrRecord } from '../../interfaces';
+import { UserProfileComponent } from '../user-profile/user-profile.component';
+import { AgoPipe } from '../../pipes/ago.pipe';
+import { DatePipe } from '@angular/common';
+import { MatTooltipModule } from '@angular/material/tooltip';
+import { LayoutService } from '../../services/layout.service';
 
 interface SocialPreview {
   url: string;
@@ -65,22 +41,25 @@ interface SocialPreview {
   selector: 'app-content',
   standalone: true,
   imports: [
+    NoteContentComponent,
     MatCardModule,
     MatProgressSpinnerModule,
+    MatTooltipModule,
     SocialPreviewComponent,
-    MatDialogModule,
+    AgoPipe,
+    DatePipe,
+    UserProfileComponent,
   ],
   templateUrl: './content.component.html',
   styleUrl: './content.component.scss',
 })
 export class ContentComponent implements AfterViewInit, OnDestroy {
-  readonly media = inject(MediaPlayerService);
-  private sanitizer = inject(DomSanitizer);
-  private dialog = inject(MatDialog);
   settings = inject(SettingsService);
   private utilities = inject(UtilitiesService);
   private router = inject(Router);
   private parsing = inject(ParsingService);
+  layoutService = inject(LayoutService);
+  data = inject(DataService);
 
   @ViewChild('contentContainer') contentContainer!: ElementRef;
   // Input for raw content
@@ -119,6 +98,10 @@ export class ContentComponent implements AfterViewInit, OnDestroy {
 
   // Social previews for URLs
   socialPreviews = signal<SocialPreview[]>([]);
+
+  eventMentions = signal<
+    { event: NostrRecord; contentTokens: ContentToken[] }[]
+  >([]);
 
   @Input() set content(value: string) {
     const newContent = value || '';
@@ -200,11 +183,32 @@ export class ContentComponent implements AfterViewInit, OnDestroy {
     this.parseDebounceTimer = window.setTimeout(async () => {
       try {
         this._isParsing.set(true);
-        const newTokens = await this.parseContent(content);
+        const newTokens = await this.parsing.parseContent(content);
+
+        const eventMentions = await Promise.all(
+          newTokens
+            .filter(
+              t => t.type === 'nostr-mention' && t.nostrData?.type === 'nevent'
+            )
+            .map(async mention => {
+              const eventData = await this.data.getEventById(
+                mention.nostrData?.data.id
+              );
+              if (!eventData) return null;
+              const contentTokens = await this.parsing.parseContent(
+                eventData?.data
+              );
+              return {
+                event: eventData,
+                contentTokens,
+              };
+            })
+        );
 
         // Use untracked to prevent triggering effects during token update
         untracked(() => {
           this._cachedTokens.set(newTokens);
+          this.eventMentions.set(eventMentions.filter(m => !!m));
           this._lastParsedContent = content;
         });
       } catch (error) {
@@ -251,376 +255,6 @@ export class ContentComponent implements AfterViewInit, OnDestroy {
 
     // Start observing the element
     this.intersectionObserver.observe(this.contentContainer.nativeElement);
-  }
-
-  private emojiMap: Record<string, string> = {
-    ':badge:': '🏅',
-    ':heart:': '❤️',
-    ':fire:': '🔥',
-    ':thumbs_up:': '👍',
-    ':thumbs_down:': '👎',
-    ':smile:': '😊',
-    ':laugh:': '😂',
-    ':cry:': '😢',
-    ':angry:': '😠',
-    ':confused:': '😕',
-    ':surprised:': '😮',
-    ':wink:': '😉',
-    ':cool:': '😎',
-    ':kiss:': '😘',
-    ':heart_eyes:': '😍',
-    ':thinking:': '🤔',
-    ':clap:': '👏',
-    ':pray:': '🙏',
-    ':muscle:': '💪',
-    ':ok_hand:': '👌',
-    ':wave:': '👋',
-    ':point_right:': '👉',
-    ':point_left:': '👈',
-    ':point_up:': '👆',
-    ':point_down:': '👇',
-    ':rocket:': '🚀',
-    ':star:': '⭐',
-    ':lightning:': '⚡',
-    ':sun:': '☀️',
-    ':moon:': '🌙',
-    ':rainbow:': '🌈',
-    ':coffee:': '☕',
-    ':beer:': '🍺',
-    ':wine:': '🍷',
-    ':pizza:': '🍕',
-    ':burger:': '🍔',
-    ':cake:': '🎂',
-    ':party:': '🎉',
-    ':gift:': '🎁',
-    ':music:': '🎵',
-    ':note:': '🎶',
-    ':phone:': '📱',
-    ':computer:': '💻',
-    ':email:': '📧',
-    ':lock:': '🔒',
-    ':unlock:': '🔓',
-    ':key:': '🔑',
-    ':money:': '💰',
-    ':dollar:': '💵',
-    ':euro:': '💶',
-    ':yen:': '💴',
-    ':pound:': '💷',
-    ':gem:': '💎',
-    ':crown:': '👑',
-    ':trophy:': '🏆',
-    ':medal:': '🏅',
-    ':first_place:': '🥇',
-    ':second_place:': '🥈',
-    ':third_place:': '🥉',
-    ':checkmark:': '✅',
-    ':cross:': '❌',
-    ':warning:': '⚠️',
-    ':stop:': '🛑',
-    ':green_circle:': '🟢',
-    ':red_circle:': '🔴',
-    ':yellow_circle:': '🟡',
-    ':blue_circle:': '🔵',
-    ':purple_circle:': '🟣',
-    ':orange_circle:': '🟠',
-    ':white_circle:': '⚪',
-    ':black_circle:': '⚫',
-  };
-
-  private async parseContent(content: string): Promise<ContentToken[]> {
-    if (!content) return [];
-
-    // Replace line breaks with placeholders
-    const processedContent = content.replace(/\n/g, '##LINEBREAK##');
-
-    // Regex for different types of content - updated to avoid capturing trailing LINEBREAK placeholders
-    // URL regex: capture potential trailing punctuation so we can trim logic-smart (e.g. parentheses, commas, periods)
-    const urlRegex =
-      /(https?:\/\/[^\s##)\]\}>]+)(?=\s|##LINEBREAK##|$|[),.;!?:])/g;
-    const youtubeRegex =
-      /(https?:\/\/)?(www\.)?(youtube\.com\/watch\?v=|youtu\.be\/)([a-zA-Z0-9_-]{11})(?=\s|##LINEBREAK##|$)/g;
-    const imageRegex =
-      /(https?:\/\/[^\s##]+\.(jpg|jpeg|png|gif|webp)(\?[^\s##]*)?(?=\s|##LINEBREAK##|$))/gi;
-    const audioRegex =
-      /(https?:\/\/[^\s##]+\.(mp3|wav|ogg)(\?[^\s##]*)?(?=\s|##LINEBREAK##|$))/gi;
-    const videoRegex =
-      /(https?:\/\/[^\s##]+\.(mp4|webm|mov|avi|wmv|flv|mkv)(\?[^\s##]*)?(?=\s|##LINEBREAK##|$))/gi;
-    const nostrRegex =
-      /(nostr:(?:npub|nprofile|note|nevent|naddr)1[a-zA-Z0-9]+)(?=\s|##LINEBREAK##|$|[^\w])/g;
-    const emojiRegex = /(:[a-zA-Z_]+:)/g;
-
-    // Split content and generate tokens
-    const tokens: ContentToken[] = [];
-    let lastIndex = 0;
-
-    // Find all matches and their positions
-    const matches: {
-      start: number;
-      end: number;
-      content: string;
-      type: ContentToken['type'];
-      nostrData?: NostrData;
-      emoji?: string;
-      processedUrl?: SafeResourceUrl;
-    }[] = [];
-
-    // Find emoji codes first (highest priority after nostr)
-    let match: RegExpExecArray | null;
-    while ((match = emojiRegex.exec(processedContent)) !== null) {
-      const emojiCode = match[0];
-      const emoji = this.emojiMap[emojiCode];
-      if (emoji) {
-        matches.push({
-          start: match.index,
-          end: match.index + match[0].length,
-          content: emojiCode,
-          type: 'emoji',
-          emoji,
-        });
-      }
-    }
-
-    // Find Nostr URIs (highest priority) - collect first, then batch process
-    const nostrMatches: {
-      match: RegExpExecArray;
-      index: number;
-      length: number;
-    }[] = [];
-    while ((match = nostrRegex.exec(processedContent)) !== null) {
-      nostrMatches.push({
-        match,
-        index: match.index,
-        length: match[0].length,
-      });
-    }
-
-    // Batch process nostr URIs to avoid sequential awaits
-    const nostrDataPromises = nostrMatches.map(async nostrMatch => {
-      try {
-        const nostrData = await this.parsing.parseNostrUri(nostrMatch.match[0]);
-        return {
-          ...nostrMatch,
-          nostrData,
-        };
-      } catch (error) {
-        console.warn('Error parsing nostr URI:', nostrMatch.match[0], error);
-        return {
-          ...nostrMatch,
-          nostrData: null,
-        };
-      }
-    });
-
-    // Wait for all nostr URIs to be processed
-    const processedNostrMatches = await Promise.all(nostrDataPromises);
-
-    // Add valid nostr matches to the matches array
-    for (const { match, index, length, nostrData } of processedNostrMatches) {
-      if (nostrData) {
-        matches.push({
-          start: index,
-          end: index + length,
-          content: match[0],
-          type: 'nostr-mention',
-          nostrData,
-        });
-      }
-    }
-
-    // Find YouTube URLs
-    while ((match = youtubeRegex.exec(processedContent)) !== null) {
-      // Pre-process the YouTube URL to avoid repeated calls in template
-      const youtubeUrl = match[0];
-      const processedUrl = this.media.getYouTubeEmbedUrl()(youtubeUrl);
-
-      matches.push({
-        start: match.index,
-        end: match.index + match[0].length,
-        content: youtubeUrl,
-        type: 'youtube',
-        processedUrl: processedUrl,
-      });
-    }
-
-    // Find image URLs
-    imageRegex.lastIndex = 0;
-    while ((match = imageRegex.exec(processedContent)) !== null) {
-      matches.push({
-        start: match.index,
-        end: match.index + match[0].length,
-        content: match[0],
-        type: 'image',
-      });
-    }
-
-    // Find video URLs
-    videoRegex.lastIndex = 0;
-    while ((match = videoRegex.exec(processedContent)) !== null) {
-      matches.push({
-        start: match.index,
-        end: match.index + match[0].length,
-        content: match[0],
-        type: 'video',
-      });
-    }
-
-    // Find audio URLs
-    audioRegex.lastIndex = 0;
-    while ((match = audioRegex.exec(processedContent)) !== null) {
-      matches.push({
-        start: match.index,
-        end: match.index + match[0].length,
-        content: match[0],
-        type: 'audio',
-      });
-    }
-
-    // Find remaining URLs
-    urlRegex.lastIndex = 0;
-    while ((match = urlRegex.exec(processedContent)) !== null) {
-      let rawUrl = match[0];
-      const start = match.index;
-
-      // Trim trailing punctuation that is unlikely part of the URL
-      const trailingPattern = /[)\],;!?.]+$/;
-      while (trailingPattern.test(rawUrl)) {
-        const lastChar = rawUrl.slice(-1);
-        if (lastChar === '/' || lastChar === '#') break; // keep structural chars
-        if (lastChar === ')') {
-          const openCount = (rawUrl.match(/\(/g) || []).length;
-          const closeCount = (rawUrl.match(/\)/g) || []).length;
-          if (closeCount <= openCount) break;
-        }
-        if (lastChar === ']') {
-          const openCount = (rawUrl.match(/\[/g) || []).length;
-          const closeCount = (rawUrl.match(/\]/g) || []).length;
-          if (closeCount <= openCount) break;
-        }
-        rawUrl = rawUrl.slice(0, -1);
-      }
-
-      if (!rawUrl) continue;
-
-      const isSpecialType = matches.some(
-        m => m.start === start && m.end === start + rawUrl.length
-      );
-      if (!isSpecialType) {
-        matches.push({
-          start,
-          end: start + rawUrl.length,
-          content: rawUrl,
-          type: 'url',
-        });
-      }
-    }
-
-    // Sort matches by their starting position
-    matches.sort((a, b) => a.start - b.start);
-
-    // Process text segments and matches with deterministic IDs
-    for (const match of matches) {
-      // Add text segment before the match
-      if (match.start > lastIndex) {
-        const textSegment = processedContent.substring(lastIndex, match.start);
-        this.processTextSegment(textSegment, tokens, lastIndex);
-      }
-
-      // Add the match as a token with deterministic ID based on position and content
-      const tokenId = this.generateStableTokenId(
-        match.start,
-        match.content,
-        match.type
-      );
-      const token: ContentToken = {
-        id: tokenId,
-        type: match.type,
-        content: match.content,
-      };
-
-      if (match.nostrData) {
-        token.nostrData = match.nostrData;
-      }
-
-      if (match.emoji) {
-        token.emoji = match.emoji;
-      }
-
-      if (match.processedUrl) {
-        token.processedUrl = match.processedUrl;
-      }
-
-      tokens.push(token);
-
-      lastIndex = match.end;
-    }
-
-    // Add remaining text after the last match
-    if (lastIndex < processedContent.length) {
-      const textSegment = processedContent.substring(lastIndex);
-      this.processTextSegment(textSegment, tokens, lastIndex);
-    }
-
-    return tokens;
-  }
-  private processTextSegment(
-    segment: string,
-    tokens: ContentToken[],
-    basePosition: number
-  ): void {
-    // Process line breaks in text segments
-    const parts = segment.split('##LINEBREAK##');
-
-    for (let i = 0; i < parts.length; i++) {
-      // Only add text token if there's actual content (not empty string)
-      if (parts[i].trim()) {
-        const tokenId = this.generateStableTokenId(
-          basePosition + i,
-          parts[i].trim(),
-          'text'
-        );
-        tokens.push({
-          id: tokenId,
-          type: 'text',
-          content: parts[i].trim(),
-        });
-      }
-
-      // Add a line break token after each part except the last one
-      if (i < parts.length - 1) {
-        const linebreakId = this.generateStableTokenId(
-          basePosition + i,
-          '',
-          'linebreak'
-        );
-        tokens.push({
-          id: linebreakId,
-          type: 'linebreak',
-          content: '',
-        });
-      }
-    }
-  }
-
-  getVideoType(url: string): string {
-    const extension = url.split('.').pop()?.split('?')[0]?.toLowerCase();
-    switch (extension) {
-      case 'mp4':
-        return 'mp4';
-      case 'webm':
-        return 'webm';
-      case 'mov':
-        return 'quicktime';
-      case 'avi':
-        return 'x-msvideo';
-      case 'wmv':
-        return 'x-ms-wmv';
-      case 'flv':
-        return 'x-flv';
-      case 'mkv':
-        return 'x-matroska';
-      default:
-        return 'mp4';
-    }
   }
 
   private async loadSocialPreviews(urls: string[]): Promise<void> {
@@ -696,21 +330,6 @@ export class ContentComponent implements AfterViewInit, OnDestroy {
     }
   }
 
-  /**
-   * Opens an image dialog to view the image with zoom capabilities
-   */
-  openImageDialog(imageUrl: string): void {
-    console.log('Opening image dialog for URL:', imageUrl);
-    this.dialog.open(ImageDialogComponent, {
-      data: { imageUrl },
-      maxWidth: '95vw',
-      maxHeight: '95vh',
-      width: '100%',
-      height: '100%',
-      panelClass: 'image-dialog',
-    });
-  }
-
   onNostrMentionClick(token: ContentToken) {
     if (!token.nostrData) return;
 
@@ -750,23 +369,4 @@ export class ContentComponent implements AfterViewInit, OnDestroy {
   shouldShowContent = computed(() => {
     return this._isVisible() || this._hasBeenVisible();
   });
-
-  /**
-   * Generate a stable token ID based on position and content
-   */
-  private generateStableTokenId(
-    position: number,
-    content: string,
-    type: string
-  ): number {
-    // Create a simple hash from position, content, and type
-    let hash = 0;
-    const str = `${position}-${type}-${content}`;
-    for (let i = 0; i < str.length; i++) {
-      const char = str.charCodeAt(i);
-      hash = (hash << 5) - hash + char;
-      hash = hash & hash; // Convert to 32-bit integer
-    }
-    return Math.abs(hash);
-  }
 }
