+++ src/app/components/user-profile/user-profile.component.scss
@@ -273,6 +273,32 @@
     }
   }
 
+  &.compact {
+    width: auto; // Changed from 100px to auto
+    flex-shrink: 0;
+
+    .user-profile-container {
+      justify-content: center; // Center the avatar
+      padding: 0; // Removed padding
+      width: auto; // Container should only be as wide as its content
+    }
+
+    .user-profile-avatar {
+      width: 24px;
+      height: 24px;
+      margin: 0; // Removed margin
+    }
+
+    .default-user-avatar {
+      font-size: 24px;
+    }
+
+    .user-profile-name {
+      font-size: 14px;
+      margin-top: 0; // Removed margin
+    }
+  }
+
   &.icon {
     width: auto; // Changed from 100px to auto
     flex-shrink: 0;
